fix: resolve build failure and consolidate MahjongCalculator (Ver 3.5.0b)

+++ lib/features/calc/calc_state.dart
@@ -479,7 +479,7 @@ class CalcNotifier extends Notifier<CalcState> {
           rate: config.rate, 
           totalChips: chipTotals[i], 
           chipRate: config.chipRate, 
-          totalFee: config.gameFee,
+          totalFee: config.gameFee.toDouble(),
           playerCount: players,
         );
       });
